refactor Navigation
