+++ frontend/components/Navigation.tsx
@@ -1,27 +1,25 @@
 import Link from "next/link";
-import EventIcon from '@mui/icons-material/Event';
-import CottageIcon from '@mui/icons-material/Cottage';
-import InfoIcon from '@mui/icons-material/Info';
+import { Event, Cottage, Info } from '@mui/icons-material';
 
-export default function Navigation () {
+export function Navigation () {
   return (
     <nav>
       <div id='navigation-links-container'>
         <div id='nav-title'>
           <Link className='nav-link' href='/'>
-            <EventIcon fontSize='large' />
+            <Event fontSize='large' />
           </Link>
         </div>
         <ul id='nav-link-container'>
           <li className='link-container'>
             <Link className='nav-link' href='/'>
-              <CottageIcon />&nbsp;
+              <Cottage />&nbsp;
               Home
             </Link>
           </li>
           <li className='link-container'>
             <Link className='nav-link' href='/about'>
-              <InfoIcon />&nbsp;
+              <Info />&nbsp;
               About
             </Link>
           </li>
