+++ frontend/components/UI/PopOverButton.tsx
@@ -1,11 +1,8 @@
 import { MouseEvent, ReactNode, useState } from 'react';
-import Popover from '@mui/material/Popover';
-import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import FavoriteIcon from '@mui/icons-material/Favorite';
-import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
+import { Button, Popover } from '@mui/material';
+import { Favorite, FavoriteBorder } from '@mui/icons-material';
 
-export default function PopoverButton({
+export function PopoverButton({
   numOfFaves,
   children
 }: {
@@ -30,10 +27,10 @@ export default function PopoverButton({
       <Button aria-describedby={id} variant={numOfFaves ? 'contained' : 'outlined'} onClick={handleClick}>
         {numOfFaves ? (
           <>
-            <FavoriteIcon style={{ color: '#d81159ff'}} />&nbsp;({numOfFaves})
+            <Favorite style={{ color: '#d81159ff'}} />&nbsp;({numOfFaves})
           </>
         ) : (
-          <FavoriteBorderIcon style={{ color: '#d81159ff'}} />
+          <FavoriteBorder style={{ color: '#d81159ff'}} />
         )}
       </Button>
       <Popover
