+++ frontend/components/UI/index.tsx
@@ -4,3 +4,4 @@ export * from '@/components/UI/ButtonGroup';
 export * from '@/components/UI/DataTable';
 export * from '@/components/UI/EventListing';
 export * from '@/components/UI/Modal';
+export * from '@/components/UI/PopOverButton';
