+++ frontend/components/index.tsx
@@ -1,1 +1,2 @@
 export * from '@/components/UI';
+export * from '@/components/Navigation';
